Complete refactorign of popup educational messages
